consolidate logic around common chunk inclusions
