+++ webpack.conf.js
@@ -5,6 +5,7 @@ var helpers = require('./gulpHelpers');
 var RequireEnsureWithoutJsonp = require('./plugins/RequireEnsureWithoutJsonp.js');
 var { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 var argv = require('yargs').argv;
+var allowedModules = require('./allowedModules');
 
 // list of module names to never include in the common bundle chunk
 var neverBundle = [
@@ -26,12 +27,14 @@ plugins.push(  // this plugin must be last so it can be easily removed for karma
     name: 'prebid',
     filename: 'prebid-core.js',
     minChunks: function(module) {
-      return (
+       return (
         (
-          module.context && module.context === path.resolve('./src') &&
+          module.context && module.context.startsWith(path.resolve('./src')) &&
           !(module.resource && neverBundle.some(name => module.resource.includes(name)))
         ) ||
-        module.resource && module.resource.includes(path.resolve('./node_modules/core-js'))
+        module.resource && (allowedModules.src.concat(['core-js'])).some(
+          name => module.resource.includes(path.resolve('./node_modules/' + name))
+        )
       );
     }
   })
